TypeTools docs
